refactor #57: Move shell completions in their own file.
